+++ CHANGELOG.md
@@ -9,3 +9,4 @@ All notable changes to Dexter will be documented in this file.
 - Add `FetchRequest.getLiquidityPoolState()` helper to get the latest state for a liquidity pool.
 - Liquidity pool fee fix for SundaeSwap when constructing pools from on-chain data. 
 - Add `SwapRequest.withSwapOutAmount(bigint)` to calculate the estimated swap in amount.
+- Add `SwapRequest.withSwapOutToken(Token)` to allow crafting a SwapRequest given the swap out token.